SD removal no longer halts the program.
Added a while loop which only breaks if the SD card is re-initialized. Also checks if the file still exists after the SD card is running. If not headers will be written to the first line.

Also added a few more Serial prints for debugging and reformatted the order of the code.

More comments also added to improve readability.